ai_lab/requirements.txt
langchain-groq
langchain-community
langchain
faiss-cpu
python-dotenv
sqlalchemy